Setting up basic structure of code
I have started the early stages of development. I have 2 stacks created, one for Operators, and the other for Operands. I started the main loop of the program, that will check each character of the infix expression. Mostly completed the multi digit parsing using a stringbuilder. Also added the isOperator method that returns a Boolean if a character is an operator.